--- a/src/Inline/Parser/BacktickParser.php
+++ b/src/Inline/Parser/BacktickParser.php
@@ -17,6 +17,11 @@ namespace League\CommonMark\Inline\Parser;
 use League\CommonMark\Inline\Element\Code;
 use League\CommonMark\Inline\Element\Text;
 use League\CommonMark\InlineParserContext;
+use function mb_substr;
+use function preg_match;
+use function preg_replace;
+use function strlen;
+use function substr;
 
 final class BacktickParser implements InlineParserInterface
 {
@@ -47,11 +52,30 @@ final class BacktickParser implements InlineParserInterface
 
         while ($matchingTicks = $cursor->match('/`+/m')) {
             if ($matchingTicks === $ticks) {
-                $code = \mb_substr($cursor->getLine(), $currentPosition, $cursor->getPosition() - $currentPosition - \strlen($ticks), 'utf-8');
-                $c = \preg_replace('/\n/m', ' ', $code);
+                if ($cursor->isMultiByte()) {
+                    $code = mb_substr(
+                        $cursor->getLine(),
+                        $currentPosition,
+                        $cursor->getPosition() - $currentPosition - strlen($ticks),
+                        $cursor->getEncoding()
+                    );
+                } else {
+                    $code = substr(
+                        $cursor->getLine(),
+                        $currentPosition,
+                        $cursor->getPosition() - $currentPosition - strlen($ticks)
+                    );
+                }
 
-                if ($c !== '' && \preg_match('/[^ ]/', $c) && \mb_substr($c, 0, 1) === ' ' && \mb_substr($c, -1, 1) === ' ') {
-                    $c = \mb_substr($c, 1, -1);
+                $c = preg_replace('/\n/m', ' ', $code);
+
+                if (
+                    false === empty($c) &&
+                    preg_match('/[^ ]/', $c) &&
+                    strpos($c, ' ') === 0 &&
+                    $c[strlen($c) - 1] === ' '
+                ) {
+                    $c = substr($c, 1, -1);
                 }
 
                 $inlineContext->getContainer()->appendChild(new Code($c));