- Updated to call multibyte methods only when necessary
- Replaced some `preg_match` calls searching for tabs with `strpos` searching for `chr(9)`
- Fixed the benchmark error "failed to close process"
- Fixed some possible bugs where it would ignore the encoding of the Cursor and just assume "UTF-8"
- Small PHP optimisations in general

This PR opens some doors as it shows some examples of performance improvements (which I'm sorry I'm not covering more as my time is limited).

With this PR I was able to be around 3~7ms faster.

Before:
```
Running Benchmarks (Isolated), 7 Implementations, 20 Iterations:
	CommonMark                     ....................
	PHP Markdown                   ....................
	PHP Markdown Extra             ....................
	Parsedown                      ....................
	cebe/markdown                  ....................
	cebe/markdown gfm              ....................
	cebe/markdown extra            ....................

Benchmark Results, CPU:
	1) Parsedown                    2.41 ms              
	2) cebe/markdown                2.92 ms      -0.51 ms
	3) cebe/markdown gfm            3.78 ms      -1.37 ms
	4) PHP Markdown                 4.23 ms      -1.82 ms
	5) cebe/markdown extra          4.70 ms      -2.29 ms
	6) PHP Markdown Extra           6.56 ms      -4.15 ms
	7) CommonMark                  31.14 ms     -28.73 ms

Benchmark Results, Peak Memory:
	1) PHP Markdown                 2413 kB              
	2) Parsedown                    2556 kB       -143 kB
	3) PHP Markdown Extra           2573 kB       -160 kB
	4) cebe/markdown                2645 kB       -232 kB
	5) cebe/markdown extra          2780 kB       -367 kB
	6) cebe/markdown gfm            3134 kB       -721 kB
	7) CommonMark                   6545 kB      -4132 kB
```

With my changes:
```
Running Benchmarks (Isolated), 7 Implementations, 20 Iterations:
	CommonMark                     ....................
	PHP Markdown                   ....................
	PHP Markdown Extra             ....................
	Parsedown                      ....................
	cebe/markdown                  ....................
	cebe/markdown gfm              ....................
	cebe/markdown extra            ....................

Benchmark Results, CPU:
	1) Parsedown                    2.45 ms              
	2) cebe/markdown                2.82 ms      -0.37 ms
	3) cebe/markdown extra          3.06 ms      -0.61 ms
	4) cebe/markdown gfm            3.86 ms      -1.41 ms
	5) PHP Markdown                 4.26 ms      -1.81 ms
	6) PHP Markdown Extra           6.47 ms      -4.02 ms
	7) CommonMark                  26.89 ms     -24.44 ms

Benchmark Results, Peak Memory:
	1) PHP Markdown                 2413 kB              
	2) Parsedown                    2556 kB       -143 kB
	3) PHP Markdown Extra           2573 kB       -160 kB
	4) cebe/markdown                2645 kB       -232 kB
	5) cebe/markdown extra          2780 kB       -367 kB
	6) cebe/markdown gfm            3134 kB       -721 kB
	7) CommonMark                   6526 kB      -4113 kB
```

This doesn't solve entirely but helps a lot #385 (at least the benchmark script problem with `-memory` flag).

Some further read that helps it:
https://mike42.me/blog/2018-06-how-i-made-my-php-code-run-100-times-faster
http://maettig.com/code/php/php-performance-benchmarks.php