@ricardofiorani 2019-10-14
I think the CI pipeline is broken but not caused by my PR as can be seen in https://travis-ci.org/thephpleague/commonmark/jobs/597589652 🤷‍♂