@colinodell 2019-10-16
Thank you for the contribution!  I'm not ignoring this - I just want to find the time to perform a solid review.  I had tried a similar approach in the past with the `Cursor` class but found that the overhead of additional `if` statements either negated the benefits or made things slower, so I'd like to take a closer look to better understand why we're seeing the positive results now (or if perhaps some individual changes are a bit slower, but they are overshadowed by significant improvements elsewhere)