@colinodell 2019-10-18
Closing this in favor of #389, which has virtually all of the same tweaks introduced here, but with a few other changes to tidy things up.

Thank you again for your contribution here!